Add conky_images.conf
Add conky_images.conf to handle drawing images to window. By separating the text and images into different files with different config settings, we are able to minimize the amount of work conky has to do. 

Since the text needs to be updated in real-time, it makes since to set the conky.conf update_interval to 1 second. 
Since the ubuntu-logo.png image is static, it makes more sense to set the conky_images.conf update_interval to 1 hour.